refactor: update ShareModal and QrCodeTab components to enhance layout and accessibility features, remove unused QrCodeDialog and related files
