+++ apps/journeys-admin/src/components/Editor/ShareModal/tabs/QrCodeTab/CodeActionButton/CodeActionButton.spec.tsx
