+++ apps/journeys-admin/src/components/Editor/ShareModal/tabs/QrCodeTab/CodeCanvas/index.ts
