+++ apps/journeys-admin/src/components/Editor/ShareModal/tabs/QrCodeTab/QrCodeTab.tsx
@@ -1,16 +1,48 @@
+import { gql, useMutation, useQuery } from '@apollo/client'
 import TabPanel from '@mui/lab/TabPanel'
 import Box from '@mui/material/Box'
-import Button from '@mui/material/Button'
+import Stack from '@mui/material/Stack'
 import type { Theme } from '@mui/material/styles'
 import Typography from '@mui/material/Typography'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import { useTranslation } from 'next-i18next'
-import { ReactElement } from 'react'
-
-import TransformIcon from '@core/shared/ui/icons/Transform'
+import { useSnackbar } from 'notistack'
+import { ReactElement, useEffect, useState } from 'react'
 
 import { GetJourneyForSharing_journey as JourneyFromLazyQuery } from '../../../../../../__generated__/GetJourneyForSharing'
+import {
+  GetJourneyQrCodes,
+  GetJourneyQrCodesVariables
+} from '../../../../../../__generated__/GetJourneyQrCodes'
 import { JourneyFields as JourneyFromContext } from '../../../../../../__generated__/JourneyFields'
+import {
+  QrCodeCreate,
+  QrCodeCreateVariables
+} from '../../../../../../__generated__/QrCodeCreate'
+import { QrCodeFields as QrCode } from '../../../../../../__generated__/QrCodeFields'
+
+import { CodeActionButton } from './CodeActionButton'
+import { CodeCanvas } from './CodeCanvas'
+import { QR_CODE_FIELDS } from './qrCodeFields'
+import { ScanCount } from './ScanCount'
+
+export const GET_JOURNEY_QR_CODES = gql`
+  ${QR_CODE_FIELDS}
+  query GetJourneyQrCodes($where: QrCodesFilter!) {
+    qrCodes(where: $where) {
+      ...QrCodeFields
+    }
+  }
+`
+
+export const QR_CODE_CREATE = gql`
+  ${QR_CODE_FIELDS}
+  mutation QrCodeCreate($input: QrCodeCreateInput!) {
+    qrCodeCreate(input: $input) {
+      ...QrCodeFields
+    }
+  }
+`
 
 interface QrCodeTabProps {
   journey?: JourneyFromContext | JourneyFromLazyQuery
@@ -18,7 +50,95 @@ interface QrCodeTabProps {
 
 export function QrCodeTab({ journey }: QrCodeTabProps): ReactElement {
   const { t } = useTranslation('apps-journeys-admin')
+  const { enqueueSnackbar } = useSnackbar()
   const mdUp = useMediaQuery((theme: Theme) => theme.breakpoints.up('md'))
+  const [qrCodeCreate, { loading: createLoading }] = useMutation<
+    QrCodeCreate,
+    QrCodeCreateVariables
+  >(QR_CODE_CREATE)
+
+  const [loading, setLoading] = useState(true)
+
+  const {
+    data,
+    loading: getLoading,
+    refetch
+  } = useQuery<GetJourneyQrCodes, GetJourneyQrCodesVariables>(
+    GET_JOURNEY_QR_CODES,
+    {
+      variables: {
+        where: {
+          journeyId: journey?.id
+        }
+      },
+      skip: journey?.id == null
+    }
+  )
+
+  useEffect(() => {
+    void refetch()
+  }, [journey?.slug, refetch])
+
+  const qrCode = data?.qrCodes[0]
+  const shortLink = getShortLink(qrCode)
+
+  useEffect(() => {
+    if (getLoading || createLoading) {
+      setLoading(true)
+    } else {
+      setLoading(false)
+    }
+  }, [getLoading, createLoading])
+
+  async function handleGenerateQrCode(): Promise<void> {
+    if (journey?.id == null || journey?.team?.id == null) return
+    await qrCodeCreate({
+      variables: {
+        input: {
+          journeyId: journey.id,
+          teamId: journey.team.id
+        }
+      },
+      update(cache, { data }) {
+        if (data?.qrCodeCreate != null) {
+          cache.modify({
+            fields: {
+              qrCodes(existingQrCodes = []) {
+                const newQrCodeRef = cache.writeFragment({
+                  data: data.qrCodeCreate,
+                  fragment: gql`
+                    fragment NewQrCode on QrCode {
+                      id
+                    }
+                  `
+                })
+                return [...existingQrCodes, newQrCodeRef]
+              }
+            }
+          })
+        }
+      },
+      onError: () => {
+        enqueueSnackbar(t('Failed to create QR Code'), {
+          variant: 'error',
+          preventDuplicate: true
+        })
+      }
+    })
+  }
+
+  function getShortLink(qrCode?: QrCode): string | undefined {
+    if (qrCode == null) return undefined
+
+    const hostname = qrCode.shortLink.domain.hostname
+    const pathname = qrCode.shortLink.pathname
+    const isLocal = hostname === 'localhost'
+
+    const protocol = isLocal ? 'http:' : 'https:'
+    const port = isLocal ? ':4100' : ''
+
+    return `${protocol}//${hostname}${port}/${pathname}`
+  }
 
   return (
     <TabPanel value="1" sx={{ padding: 0 }}>
@@ -39,33 +159,34 @@ export function QrCodeTab({ journey }: QrCodeTabProps): ReactElement {
           </Typography>
         </Box>
 
-        <Box
+        <Stack
           sx={{
-            display: 'flex',
-            flexDirection: 'column',
+            flexDirection: { xs: 'column', sm: 'row' },
+            gap: { xs: 4, sm: 7 },
             alignItems: 'center',
-            gap: 2,
-            py: 4,
-            border: '2px dashed',
-            borderColor: 'divider',
-            borderRadius: 2,
-            backgroundColor: 'background.paper'
+            py: 2
           }}
         >
-          <TransformIcon sx={{ fontSize: 48, color: 'text.secondary' }} />
-          <Typography variant="body2" color="text.secondary" align="center">
-            {t('Click the button below to generate a QR code for your journey')}
-          </Typography>
-          <Button
-            variant="contained"
-            color="primary"
-            startIcon={<TransformIcon />}
-            disabled={journey == null}
-            sx={{ mt: 1 }}
+          <CodeCanvas shortLink={shortLink} loading={loading} />
+          <Stack
+            spacing={3}
+            sx={{
+              alignItems: { xs: 'center', sm: 'start' }
+            }}
           >
-            {t('Generate QR Code')}
-          </Button>
-        </Box>
+            <ScanCount shortLinkId={qrCode?.shortLink.id} />
+            <CodeActionButton
+              shortLink={shortLink}
+              loading={loading}
+              handleGenerateQrCode={handleGenerateQrCode}
+            />
+            <Typography variant="body2" color="secondary.main">
+              {t(
+                'Here is the unique QR code for your Journey. You can change the Journey URL without needing to re-generate the QR code'
+              )}
+            </Typography>
+          </Stack>
+        </Stack>
       </Box>
     </TabPanel>
   )
