--- apps/journeys-admin/src/components/Editor/Toolbar/Items/ShareItem/QrCodeDialog/index.ts
@@ -1,1 +1,0 @@
-export { QrCodeDialog } from './QrCodeDialog'
